Split tests into their own folder
