+++ src/BaseFragmentFactory.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const assert = require('./assert');
 
 /**
