+++ src/Dot.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const doT = require('dot'); // eslint-disable-line no-unused-vars
 const i18n = require('./i18n');
