+++ src/PubSubHub.js
@@ -1,5 +1,3 @@
-'use strict';
-
 /**
  * Implements a simple 2 way stream of events
  *
