+++ src/TextInput/index.js
@@ -1,5 +1,3 @@
-'use strict';
-
 require('./styles.css');
 const inputTmpl = require('./input.tmpl');
 const BaseTextInput = require('./BaseTextInput');
