+++ src/assert.js
@@ -1,5 +1,3 @@
-'use strict';
-
 module.exports = (condition, message) => {
   if (!condition) {
     throw new Error(message || 'Assertion failure');
