+++ src/i18n.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const i18next = require('i18next');
 
 let instance = null;
