+++ src/init.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const i18n = require('./i18n');
 
 module.exports = function init(opts, cb) {
