+++ src/juice.js
@@ -1,5 +1,3 @@
-'use strict';
-
 require('imports?jQuery=cash-dom!../vendor/jquery.simplePagination.js');
 
 require('babel-polyfill');
