+++ tests.webpack.js
@@ -1,7 +1,5 @@
-'use strict';
-
 // Tests
-const testsContext = require.context('./src', true, /-test\.js$/);
+const testsContext = require.context('./tests', true, /-test\.js$/);
 testsContext.keys().forEach(testsContext);
 
 // Components
