+++ tests/singleSelect-test.js
@@ -1,6 +1,4 @@
-'use strict';
-
-const SingleSelect = require('./index.js');
+const SingleSelect = require('../src/SingleSelect/index.js');
 
 describe('SingleSelect functionality with simple values', () => {
   let categories;
