+++ tests/spinner-test.js
@@ -1,7 +1,6 @@
 /* eslint-disable lines-around-comment */
-'use strict';
 
-const Spinner = require('./index.js');
+const Spinner = require('../src/Spinner/index.js');
 
 describe('Spinner on Window', () => {
   let spinner;
@@ -33,8 +32,9 @@ describe('Spinner on Window', () => {
     spinner.start();
 
     setTimeout(() => {
-      const overlayZindex = $('.juicy-spinner-container').css('z-index');
-      const buttonZindex = $('.ui-button').css('z-index');
+      debugger;
+      const overlayZindex = $('.juicy-spinner-container').css('zIndex');
+      const buttonZindex = $('.ui-button').css('zIndex');
 
       expect($('body')).toHaveClass('noScroll');
       expect(overlayZindex).toBe('9999');
